[Refactor] domain class PK Long 래퍼 클래스로 변경, LocalDateTime NotEmpty 삭제

+++ src/main/java/project/moseup/domain/BankBook.java
@@ -22,7 +22,7 @@ public class BankBook {
 
 	@Id @GeneratedValue
 	@Column(name = "bankbook_dealno")
-	private int dno;
+	private Long dno;
 	
 	@OneToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "member_no")
